cargo_service/requirements.txt
Django
djangorestframework
psycopg2
geopy
